backend/src/models/schemas/common.py
"""
通用 Schema
"""
from typing import Generic, TypeVar, Optional
from pydantic import BaseModel, Field


T = TypeVar("T")


class ResponseSchema(BaseModel, Generic[T]):
    """统一响应格式"""

    code: int = Field(default=0, description="状态码 0表示成功")
    message: str = Field(default="success", description="响应消息")
    data: Optional[T] = Field(default=None, description="响应数据")


class PaginationSchema(BaseModel):
    """分页参数"""

    page: int = Field(default=1, ge=1, description="页码")
    page_size: int = Field(default=20, ge=1, le=100, description="每页数量")


class PaginatedResponseSchema(BaseModel, Generic[T]):
    """分页响应"""

    items: list[T] = Field(description="数据列表")
    total: int = Field(description="总数")
    page: int = Field(description="当前页码")
    page_size: int = Field(description="每页数量")
