backend/pyproject.toml
[project]
name = "spot-backend"
version = "0.1.0"
description = "义城上门教育 - 后端服务"
requires-python = ">=3.12"
dependencies = [
    "asyncmy>=0.2.11",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.4",
    "fastapi>=0.128.0",
    "greenlet>=3.3.1",
    "loguru>=0.7.3",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
